1.2.0 - update
Wordy Fresh 1.2.0  - это масштабное обновление для всей системы и почти всех страниц.

На страницу игры добавлены:
 • Подсказки уже угаданных букв
 • Возможность использовать до 2 подсказок случайных букв за раунд!
 • Добавлена функция долгого нажатия на Backspace для поочередного удаления символов со строки вместо нескольких нажатий
 • Добавлено уведомление (с возможностью отключения) о недоигранной партии в начале новой игры
• Добавлена вибрация для нажатия клавиш, эффектов при победе или использовании других клавиш (с возможностью отключения)

Также масштабное изменение коснулось страницы статистики, где появились новые метрики для анализа своей игры! Теперь можно смотреть распределение своих игр по языку и длине загаданного слова!

Кроме этого теперь просмотр своего профиля доступен в оффлайн-режиме без возможности выхода и редактирования!

На прочих страницах:
• Добавлены переходы между текстовыми полями в формах, связанных с профилем
• Улучшена логика обработки ошибок
• Добавлена локализация на русский и английский для ошибок
• Добавлена иконка приложения в загрузочный экран приложения
• Обновлено обучение (перейдите в Настройки и ознакомьтесь!)

+++ app/src/main/java/com/sinya/projects/wordle/domain/useCase/SignOutUseCase.kt
@@ -1,6 +1,8 @@
 package com.sinya.projects.wordle.domain.useCase
 
 import android.util.Log
+import com.sinya.projects.wordle.domain.checker.NetworkChecker
+import com.sinya.projects.wordle.domain.error.NoInternetException
 import com.sinya.projects.wordle.domain.repository.AchievementRepository
 import com.sinya.projects.wordle.domain.repository.AvatarRepository
 import com.sinya.projects.wordle.domain.repository.DictionaryRepository
@@ -17,22 +19,22 @@ class SignOutUseCase @Inject constructor(
     private val dictionaryRepository: DictionaryRepository,
     private val achievementsRepository: AchievementRepository,
     private val profileRepository: ProfileRepository,
-    private val avatarRepository: AvatarRepository
+    private val avatarRepository: AvatarRepository,
+    private val networkChecker: NetworkChecker
 ) {
     suspend operator fun invoke(): Result<Unit> = withContext(Dispatchers.IO) {
         try {
-            val user = authRepository.getCurrentUser()
-                ?: return@withContext Result.success(Unit) // Уже вышли
+            if (!networkChecker.isInternetAvailable()) return@withContext Result.failure(NoInternetException())
 
-            // 1. Синхронизируем все данные в Supabase перед выходом
+            val user = authRepository.getCurrentUser()
+                ?: return@withContext Result.success(Unit)
+
             statisticsRepository.syncFromLocal().getOrThrow()
             dictionaryRepository.syncFromLocal().getOrThrow()
             achievementsRepository.syncFromLocal().getOrThrow()
 
-            // 2. Удаляем локальный аватар
             avatarRepository.deleteLocalAvatar(user.id)
 
-            // 3. Очищаем все локальные данные
             statisticsRepository.clearLocal()
             dictionaryRepository.clearLocal()
             achievementsRepository.clearLocal()
